requirements.txt
# pipenv install -r requirements.txt

wheel==0.38.1
gunicorn==20.0.4

psycopg2-binary==2.8.6

django==3.2.18
djangorestframework==3.12.2
django-countries==7.3.2
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
django-debug-toolbar==2.2.1
django-debug-toolbar-user-switcher==2.2.0
django-webpack4-loader==0.0.5
python-decouple==3.3
django-import-export==2.7.1
django-extensions==3.1.1

phonenumbers==8.12.4
pillow==9.3.0

pre-commit==3.0.2
black==22.3.0
beautifulsoup4==4.10.0

factory_boy==3.1.0
pytest==7.1.2
pytest-django==4.1.0
pytest-xdist==2.2.1
pytest-cov==2.11.1
tblib==1.7.0
